Added RiderList information and fixed float in amount for Rider to float instead of having it as int. Also edited in the class diagram.

+++ AOOAD_Ass/Policies/Rider.cs
@@ -10,9 +10,9 @@ namespace AOOAD_Ass.Policies
     {
         public int riderNo { get; set; }
         public string riderType { get; set; }
-        public int amount { get; set; }
+        public float amount { get; set; }
         public Rider() { }
-        public Rider(int no, string type, int a)
+        public Rider(int no, string type, float a)
         {
             riderNo = no;
             riderType = type;
